Togle for nightmode

+++ src/main/java/com/codecool/travelhelper/API/rapidapi/model/apimodel/WeatherDtoRapidApi.java
@@ -9,8 +9,8 @@ public class WeatherDtoRapidApi {
     private String mainParameter;
     private String description;
     private String icon;
-    private float temperature;
-    private float feelsLike;
+    private int temperature;
+    private int feelsLike;
     private int pressure;
     private int humidity;
     private float wingSpeed;
